requirements.txt
Flask==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
Flask-Cors==3.0.10
python-dotenv==0.19.2
